remove inputcontroller and remove xbox controller support
