--- core/src/com/fallenflame/game/InputController.java
@@ -1,291 +1,0 @@
-/*
- * InputController.java
- *
- * This class buffers in input from the devices and converts it into its
- * semantic meaning. If your game had an option that allows the player to
- * remap the control keys, you would store this information in this class.
- * That way, the main GameEngine does not have to keep track of the current
- * key mapping.
- *
- * Author: Walker M. White
- * Based on original PhysicsDemo Lab by Don Holden, 2007
- * B2Lights version, 3/12/2016
- */
-package com.fallenflame.game;
-import com.badlogic.gdx.*;
-
-import com.badlogic.gdx.math.Vector2;
-import com.fallenflame.game.util.XBox360Controller;
-
-/**
- * Class for reading player input.
- *
- * This supports both a keyboard and X-Box controller. In previous solutions, we only
- * detected the X-Box controller on start-up.  This class allows us to hot-swap in
- * a controller via the new XBox360Controller class.
- * @author: Walker M. White
- */
-public class InputController {
-    /** The singleton instance of the input controller */
-    private static InputController theController = null;
-
-    /**
-     * Return the singleton instance of the input controller
-     *
-     * @return the singleton instance of the input controller
-     */
-    public static InputController getInstance() {
-        if (theController == null) {
-            theController = new InputController();
-        }
-        return theController;
-    }
-
-    // Fields to manage buttons
-    /** Whether the reset button was pressed. */
-    private boolean resetPressed;
-    private boolean resetPrevious;
-    /** Whether the debug toggle was pressed. */
-    private boolean debugPressed;
-    private boolean debugPrevious;
-    /** Whether the debug2 toggle was pressed. */
-    private boolean debug2Pressed;
-    private boolean debug2Previous;
-    /** Whether the exit button was pressed. */
-    private boolean exitPressed;
-    private boolean exitPrevious;
-    /** Whether the flare button was pressed. */
-    private boolean flarePressed;
-    private boolean flarePrevious;
-    /** Whether the sprint button was pressed. */
-    private boolean sprintPressed;
-    private boolean sprintPrevious;
-    /** Whether the sneak button was pressed. */
-    private boolean sneakPressed;
-    private boolean sneakPrevious;
-
-
-    /** How much did we move horizontally? */
-    private float horizontal;
-    /** How much did we move vertically? */
-    private float vertical;
-
-    /** Did we increase or decrease the radius?
-     * Can only be two values: -1 and 1
-     */
-    private float lightRadius;
-
-    /** An X-Box controller (if it is connected) */
-    XBox360Controller xbox;
-
-    /**
-     * Returns the amount of sideways movement.
-     *
-     * -1 = left, 1 = right, 0 = still
-     *
-     * @return the amount of sideways movement.
-     */
-    public float getHorizontal() {
-        return horizontal;
-    }
-
-    /**
-     * Returns the amount of vertical movement.
-     *
-     * -1 = down, 1 = up, 0 = still
-     *
-     * @return the amount of vertical movement.
-     */
-    public float getVertical() {
-        return vertical;
-    }
-    public float getLightRadius(){
-        return lightRadius;
-    }
-
-    public Vector2 getMousePosition(){
-
-      return new Vector2(Gdx.input.getX(), Gdx.input.getY());
-
-    }
-
-    /**
-     * Returns true if the reset button was pressed.
-     *
-     * @return true if the reset button was pressed.
-     */
-    public boolean didReset() {
-        return resetPressed && !resetPrevious;
-    }
-
-    /**
-     * Returns true if the player wants to go toggle the debug mode.
-     *
-     * @return true if the player wants to go toggle the debug mode.
-     */
-    public boolean didDebug() {
-        return debugPressed && !debugPrevious;
-    }
-
-    /**
-     * Returns true if the player wants to go toggle the debug 2 mode.
-     *
-     * @return true if the player wants to go toggle the debug 2 mode.
-     */
-    public boolean didDebug2() {
-        return debug2Pressed && !debug2Previous;
-    }
-
-    /**
-     * Returns true if the exit button was pressed.
-     *
-     * @return true if the exit button was pressed.
-     */
-    public boolean didExit() {
-        return exitPressed && !exitPrevious;
-    }
-
-    /**
-     * Returns true if the fire button was pressed.
-     *
-     * @return true if the fire button was pressed.
-     */
-    public boolean didFlare() {
-        return flarePressed && !flarePrevious;
-    }
-
-    public boolean didLight() {return lightRadius != 0.0f;}
-
-    /**
-     * Returns true if the sprint button was pressed after just being not pressed.
-     *
-     * @return true if the sprint button was pressed after just being not pressed.
-     */
-    public boolean didStartSprint() { return sprintPressed && !sprintPrevious; }
-
-    /**
-     * Returns true if the sprint button was not pressed after just being pressed.
-     *
-     * @return true if the sprint button was not pressed after just being pressed.
-     */
-    public boolean didEndSprint() { return !sprintPressed && sprintPrevious; }
-
-    /**
-     * Returns true if the sneak button was pressed after just being not pressed.
-     *
-     * @return true if the sneak button was pressed after just being not pressed.
-     */
-    public boolean didStartSneak() { return sneakPressed && !sneakPrevious; }
-
-    /**
-     * Returns true if the sneak button was not pressed after just being pressed.
-     *
-     * @return true if the sneak button was not pressed after just being pressed.
-     */
-    public boolean didEndSneak() { return !sneakPressed && sneakPrevious; }
-
-    /**
-     * Creates a new input controller
-     *
-     * The input controller attempts to connect to the X-Box controller at device 0,
-     * if it exists.  Otherwise, it falls back to the keyboard control.
-     */
-    public InputController() {
-        // If we have a game-pad for id, then use it.
-        xbox = new XBox360Controller(0);
-    }
-
-    /**
-     * Reads the input for the player and converts the result into game logic.
-     */
-    public void readInput() {
-        // Copy state from last animation frame
-        // Helps us ignore buttons that are held down
-        resetPrevious  = resetPressed;
-        debugPrevious  = debugPressed;
-        debug2Previous = debug2Pressed;
-        exitPrevious = exitPressed;
-        flarePrevious = flarePressed;
-        sprintPrevious = sprintPressed;
-        sneakPrevious = sneakPressed;
-
-        // Check to see if a GamePad is connected
-        if (xbox.isConnected()) {
-            readGamepad();
-            readKeyboard(true); // Read as a back-up
-        } else {
-            readKeyboard(false);
-        }
-    }
-
-    /**
-     * TODO: if we actually want to support controller, we need to review these controls (especially how to aim flares)
-     *
-     * Reads input from an X-Box controller connected to this computer.
-     *
-     * The method provides both the input bounds and the drawing scale.  It needs
-     * the drawing scale to convert screen coordinates to world coordinates.  The
-     * bounds are for the crosshair.  They cannot go outside of this zone.
-
-     */
-    private void readGamepad() {
-        resetPressed = xbox.getStart();
-        exitPressed  = xbox.getBack();
-        debugPressed  = xbox.getY();
-        debug2Pressed = xbox.getX();
-        flarePressed = xbox.getB();
-        sprintPressed = xbox.getA();
-        sneakPressed = xbox.getRB();
-        // Increase animation frame, but only if trying to move
-        horizontal = xbox.getLeftX();
-        vertical   = xbox.getLeftY();
-        lightRadius = xbox.getLeftTrigger() == 1.0f ? 1.0f : xbox.getRightTrigger() == 1.0f ? -1.0f : 0;
-    }
-
-    /**
-     * Reads input from the keyboard.
-     *
-     * This controller reads from the keyboard regardless of whether or not an X-Box
-     * controller is connected.  However, if a controller is connected, this method
-     * gives priority to the X-Box controller.
-     *
-     * @param secondary true if the keyboard should give priority to a gamepad
-     */
-    private void readKeyboard(boolean secondary) {
-        // Give priority to gamepad results
-        resetPressed = (secondary && resetPressed) || (Gdx.input.isKeyPressed(Input.Keys.R));
-        debugPressed = (secondary && debugPressed) || (Gdx.input.isKeyPressed(Input.Keys.G));
-        debug2Pressed = (secondary && debug2Pressed) || (Gdx.input.isKeyPressed(Input.Keys.E));
-        exitPressed  = (secondary && exitPressed) || (Gdx.input.isKeyPressed(Input.Keys.ESCAPE));
-        flarePressed  = (secondary && flarePressed) || (Gdx.input.isButtonPressed(Input.Buttons.LEFT));
-        sprintPressed = (secondary && sprintPressed) || (Gdx.input.isKeyPressed(Input.Keys.SHIFT_LEFT));
-        sneakPressed = (secondary && sneakPressed) || (Gdx.input.isKeyPressed(Input.Keys.CONTROL_LEFT));
-
-        // Directional controls
-        horizontal = (secondary ? horizontal : 0.0f);
-        if (Gdx.input.isKeyPressed(Input.Keys.D)) {
-            horizontal += 1.0f;
-        }
-        if (Gdx.input.isKeyPressed(Input.Keys.A)) {
-            horizontal -= 1.0f;
-        }
-
-        // TODO: REMOVE CODE BELOW WHEN MOUSE WHEEL IS FIXED --> We should still include this as a functional alternative if possible
-        //#region mouse wheel alternative
-        if(Gdx.input.isKeyPressed(Input.Keys.PERIOD)){
-            lightRadius += 1.0f;
-        }
-        if(Gdx.input.isKeyPressed(Input.Keys.COMMA)){
-            lightRadius += -1.0f;
-        }
-        //#endregion
-
-        vertical = (secondary ? vertical : 0.0f);
-        if (Gdx.input.isKeyPressed(Input.Keys.W)) {
-            vertical += 1.0f;
-        }
-        if (Gdx.input.isKeyPressed(Input.Keys.S)) {
-            vertical -= 1.0f;
-        }
-    }
-}
